refactor:login 코드 리뷰 사항 수정
