--- src/components/Login/LoginButton/loginButtonObject.tsx
@@ -1,18 +1,0 @@
-import Apple from "./svgs/apple.svg";
-import Google from "./svgs/google.svg";
-import Kakao from "./svgs/kakao.svg";
-
-export const loginButtonObject = {
-  apple: {
-    icon: <Apple />,
-    text: "Apple로 계속하기",
-  },
-  google: {
-    icon: <Google />,
-    text: "구글로 계속하기",
-  },
-  kakao: {
-    icon: <Kakao />,
-    text: "카카오톡으로 계속하기",
-  },
-};
